Remove unnecessary TException.message hack
Thrift attempts to work-around the Python 2.7 DeprecationWarning
around `BaseException.message` by using a property.  This hack is
unnecessary since `message` is _also_ written as a regular attribute in
the `TException` constructor (and would be in any of its children),
hence the `BaseException_get_message()` wouldn't even be called.

This hack also stands in the way of making exception instances
immutable (which is a prerequisute to fixing THRIFT-4002).